+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ## [Unreleased] - ReleaseDate
 
+### Added
+
+- Add option to chain values from response header rather than body ([#184](https://github.com/LucasPickering/slumber/issues/184))
+
 ### Changed
 
 - Reduce UI latency under certain scenarios